requirements.txt
fastapi==0.117.1
uvicorn==0.35.0
torch==2.5.1
torchvision==0.20.1
pillow==11.2.1
opencv-python-headless==4.11.0.86
numpy==2.2.4
python-multipart==0.0.20
